мненого изменил структуру экшионов и редьюсеров
