+++ src/components/my-list/my-list.jsx
@@ -38,8 +38,8 @@ MyList.propTypes = {
   onMovieCardClick: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = ({data}) => ({
-  movies: data.movies,
+const mapStateToProps = ({load}) => ({
+  movies: load.movies,
 });
