+++ src/store/actions/movies.js
@@ -11,21 +11,6 @@ export const showMoreMovies = (visibleMoviesCount) => ({
   visibleMoviesCount,
 });
 
-export const loadMovies = (movies) => ({
-  type: ActionType.LOAD_MOVIES,
-  movies,
-});
-
-export const loadMoviePromo = (moviePromo) => ({
-  type: ActionType.LOAD_MOVIE_PROMO,
-  moviePromo,
-});
-
 export const enableApplication = () => ({
   type: ActionType.ENABLE_APPLICATION,
-});
-
-export const enableAuth = (status) => ({
-  type: ActionType.ENABLE_AUTH,
-  status
 });
