+++ src/store/actions/user.js
@@ -1,0 +1,7 @@
+import {ActionType} from 'const';
+
+
+export const enableAuth = (status) => ({
+  type: ActionType.ENABLE_AUTH,
+  status
+});
